refactor: config, common 패키지 하나로 모음
