+++ src/main/java/mafia/mafiatogether/chat/application/ChatService.java
@@ -8,8 +8,8 @@ import mafia.mafiatogether.chat.application.dto.response.ChatResponse;
 import mafia.mafiatogether.chat.domain.Chat;
 import mafia.mafiatogether.chat.domain.ChatRepository;
 import mafia.mafiatogether.chat.domain.Message;
-import mafia.mafiatogether.config.exception.ExceptionCode;
-import mafia.mafiatogether.config.exception.GameException;
+import mafia.mafiatogether.common.exception.ExceptionCode;
+import mafia.mafiatogether.common.exception.GameException;
 import mafia.mafiatogether.job.domain.PlayerJob;
 import mafia.mafiatogether.job.domain.PlayerJobRepository;
 import org.springframework.stereotype.Service;
