+++ src/main/java/mafia/mafiatogether/chat/ui/ChatV2Controller.java
@@ -6,8 +6,8 @@ import mafia.mafiatogether.chat.application.ChatV2Service;
 import mafia.mafiatogether.chat.application.dto.ChatV2Response;
 import mafia.mafiatogether.chat.application.dto.request.ChatRequest;
 import mafia.mafiatogether.chat.domain.Message;
-import mafia.mafiatogether.config.PlayerInfo;
-import mafia.mafiatogether.config.PlayerInfoDto;
+import mafia.mafiatogether.common.annotation.PlayerInfo;
+import mafia.mafiatogether.common.resolver.PlayerInfoDto;
 import org.springframework.http.ResponseEntity;
 import org.springframework.messaging.handler.annotation.DestinationVariable;
 import org.springframework.messaging.handler.annotation.MessageMapping;
