+++ src/main/java/mafia/mafiatogether/common/config/SwaggerConfig.java
@@ -1,4 +1,4 @@
-package mafia.mafiatogether.config;
+package mafia.mafiatogether.common.config;
 
 import static org.springframework.http.HttpHeaders.AUTHORIZATION;
 
@@ -9,6 +9,8 @@ import io.swagger.v3.oas.models.OpenAPI;
 import io.swagger.v3.oas.models.info.Info;
 import io.swagger.v3.oas.models.security.SecurityRequirement;
 import java.util.List;
+
+import mafia.mafiatogether.common.annotation.PlayerInfo;
 import org.springdoc.core.utils.SpringDocUtils;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
