+++ src/main/java/mafia/mafiatogether/common/exception/AuthException.java
@@ -1,4 +1,4 @@
-package mafia.mafiatogether.config.exception;
+package mafia.mafiatogether.common.exception;
 
 public class AuthException extends GlobalException{
     public AuthException(final ExceptionCode code) {
