+++ src/main/java/mafia/mafiatogether/common/exception/ErrorResponse.java
@@ -1,4 +1,4 @@
-package mafia.mafiatogether.config.exception;
+package mafia.mafiatogether.common.exception;
 
 import java.time.LocalDateTime;
 import java.util.ArrayList;
