+++ src/main/java/mafia/mafiatogether/common/exception/GameException.java
@@ -1,4 +1,4 @@
-package mafia.mafiatogether.config.exception;
+package mafia.mafiatogether.common.exception;
 
 public class GameException extends GlobalException{
