+++ src/main/java/mafia/mafiatogether/common/exception/GlobalException.java
@@ -1,4 +1,4 @@
-package mafia.mafiatogether.config.exception;
+package mafia.mafiatogether.common.exception;
 
 import lombok.Getter;
