+++ src/main/java/mafia/mafiatogether/common/exception/PlayerException.java
@@ -1,4 +1,4 @@
-package mafia.mafiatogether.config.exception;
+package mafia.mafiatogether.common.exception;
 
 public class PlayerException extends GlobalException{
     public PlayerException(final ExceptionCode code) {
