+++ src/main/java/mafia/mafiatogether/common/resolver/PlayerArgumentResolver.java
@@ -1,10 +1,11 @@
-package mafia.mafiatogether.config;
+package mafia.mafiatogether.common.resolver;
 
 import jakarta.servlet.http.HttpServletRequest;
 
-import mafia.mafiatogether.common.AuthExtractor;
-import mafia.mafiatogether.config.exception.AuthException;
-import mafia.mafiatogether.config.exception.ExceptionCode;
+import mafia.mafiatogether.common.util.AuthExtractor;
+import mafia.mafiatogether.common.annotation.PlayerInfo;
+import mafia.mafiatogether.common.exception.AuthException;
+import mafia.mafiatogether.common.exception.ExceptionCode;
 import org.springframework.core.MethodParameter;
 import org.springframework.web.bind.support.WebDataBinderFactory;
 import org.springframework.web.context.request.NativeWebRequest;
