+++ src/main/java/mafia/mafiatogether/game/application/GameService.java
@@ -4,8 +4,8 @@ import java.io.IOException;
 import java.time.Clock;
 import java.util.Optional;
 import lombok.RequiredArgsConstructor;
-import mafia.mafiatogether.config.exception.ExceptionCode;
-import mafia.mafiatogether.config.exception.GameException;
+import mafia.mafiatogether.common.exception.ExceptionCode;
+import mafia.mafiatogether.common.exception.GameException;
 import mafia.mafiatogether.game.application.dto.response.GameInfoResponse;
 import mafia.mafiatogether.game.application.dto.response.GameResultResponse;
 import mafia.mafiatogether.game.application.dto.response.GameStatusResponse;
