+++ src/main/java/mafia/mafiatogether/game/domain/PlayerCollection.java
@@ -4,8 +4,8 @@ import java.util.ArrayList;
 import java.util.List;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
-import mafia.mafiatogether.config.exception.ExceptionCode;
-import mafia.mafiatogether.config.exception.PlayerException;
+import mafia.mafiatogether.common.exception.ExceptionCode;
+import mafia.mafiatogether.common.exception.PlayerException;
 import mafia.mafiatogether.job.domain.jobtype.JobType;
 import mafia.mafiatogether.lobby.domain.Participant;
 import mafia.mafiatogether.lobby.domain.ParticipantCollection;
