+++ src/main/java/mafia/mafiatogether/game/domain/status/DeletedStatus.java
@@ -1,8 +1,8 @@
 package mafia.mafiatogether.game.domain.status;
 
 import lombok.NoArgsConstructor;
-import mafia.mafiatogether.config.exception.ExceptionCode;
-import mafia.mafiatogether.config.exception.GameException;
+import mafia.mafiatogether.common.exception.ExceptionCode;
+import mafia.mafiatogether.common.exception.GameException;
 import mafia.mafiatogether.game.domain.Game;
 
 @NoArgsConstructor
