+++ src/main/java/mafia/mafiatogether/lobby/domain/LobbyInfo.java
@@ -7,8 +7,8 @@ import java.util.List;
 import java.util.Queue;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
-import mafia.mafiatogether.config.exception.ExceptionCode;
-import mafia.mafiatogether.config.exception.GameException;
+import mafia.mafiatogether.common.exception.ExceptionCode;
+import mafia.mafiatogether.common.exception.GameException;
 import mafia.mafiatogether.job.domain.jobtype.Doctor;
 import mafia.mafiatogether.job.domain.jobtype.Job;
 import mafia.mafiatogether.job.domain.jobtype.Mafia;
