+++ src/test/java/mafia/mafiatogether/game/ui/GameControllerTest.java
@@ -7,8 +7,8 @@ import io.restassured.RestAssured;
 import io.restassured.http.ContentType;
 import java.util.Base64;
 import java.util.Map;
-import mafia.mafiatogether.config.exception.ErrorResponse;
-import mafia.mafiatogether.config.exception.ExceptionCode;
+import mafia.mafiatogether.common.exception.ErrorResponse;
+import mafia.mafiatogether.common.exception.ExceptionCode;
 import mafia.mafiatogether.game.application.dto.response.GameInfoResponse;
 import mafia.mafiatogether.game.application.dto.response.GameStatusResponse;
 import mafia.mafiatogether.game.application.dto.response.PlayerResponse;
