+++ src/test/java/mafia/mafiatogether/lobby/ui/LobbyControllerTest.java
@@ -5,8 +5,8 @@ import io.restassured.http.ContentType;
 import java.util.Base64;
 import java.util.List;
 import java.util.stream.Stream;
-import mafia.mafiatogether.config.exception.ErrorResponse;
-import mafia.mafiatogether.config.exception.ExceptionCode;
+import mafia.mafiatogether.common.exception.ErrorResponse;
+import mafia.mafiatogether.common.exception.ExceptionCode;
 import mafia.mafiatogether.global.ControllerTest;
 import mafia.mafiatogether.lobby.application.dto.request.LobbyCreateRequest;
 import mafia.mafiatogether.lobby.application.dto.response.LobbyCodeResponse;
